requirements.txt
astropy
gwtm_api==0.1.6
beautifulsoup4
bs4
email-validator
ephem
Flask==2.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==2.5.1
Flask-WTF
flask-caching
GeoAlchemy2>=0.13.0
healpy
pyjwt
matplotlib
numpy
pandas
plotly
psycopg2-binary
pytz
requests
scipy
Shapely
SQLAlchemy==1.4.32
urllib3
validators
Werkzeug==2.0.3
WTForms
WTForms-Components
gunicorn
boto3==1.26.90
pygcn
MOCpy
redis
emails
gcn-kafka
fsspec
adlfs
s3fs==0.6.0
markupsafe>=2.1.1
flask_profiler==1.8.1
